Remove not needed vuepress plugins
